Fixed statistics display
