--- Wordle.Api/Wordle.Api/Controllers/PlayController.cs
@@ -1,27 +1,0 @@
-﻿using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Wordle.Api.Data;
-using Wordle.Api.Dtos;
-using Wordle.Api.Services;
-using static System.Runtime.InteropServices.JavaScript.JSType;
-
-namespace Wordle.Api.Controllers
-{
-    [Route("[controller]")]
-    [ApiController]
-    public class PlayController : ControllerBase
-    {
-        private readonly PlayService _PlayService;
-        public PlayController(PlayService playService)
-        {
-            _PlayService = playService;
-        }
-
-        [HttpGet("GetLastTenDates")]
-        public async Task<List<WordOfTheDayStatsDto>> GetLastTenDates()
-        {
-            return await _PlayService.GetWordOfTheDayStats();
-        }
-
-    }
-}
