+++ Wordle.Api/Wordle.Api/Program.cs
@@ -36,7 +36,6 @@ builder.Services.AddDbContext<AppDbContext>(options =>
 });
 builder.Services.AddScoped<WordService>();
 builder.Services.AddScoped<PlayerService>();
-builder.Services.AddScoped<PlayService>();
 
 // Actually build the app so we can configure the pipeline next
 var app = builder.Build();
