+++ Wordle.Api/Wordle.Api/Services/WordService.cs
@@ -112,5 +112,36 @@ namespace Wordle.Api.Services
                 }
             }
         }
+
+        public async Task<List<WordOfTheDayStatsDto>> GetWordOfTheDayStats(DateTime? date = null, int days = 10, Guid? playerId = null)
+        {
+            if (days < 1 || days > 100) days = 10;
+            var start = date ?? DateTime.UtcNow.AddHours(-12).Date;
+            var end = start + TimeSpan.FromDays(days * -1);
+
+            var result = await _db.DateWords
+                .Include(d => d.Plays)
+                .Where(d => d.Date <= start && d.Date > end)
+                .OrderByDescending(d => d.Date)
+                .Select(d => new WordOfTheDayStatsDto
+                {
+                    Date = d.Date,
+                    AvgTime = d.Plays.Any() ? d.Plays.Average(a => a.DurationInSeconds) : -1,
+                    AvgAttempts = d.Plays.Any() ? d.Plays.Average(a => a.Attempts) : -1,
+                    NumberOfPlays = d.Plays.Count(),
+                    HasUserPlayed = playerId.HasValue && d.Plays.Any(d => d.PlayerId == playerId.Value)
+                })
+                .ToListAsync();
+
+            if (result.Count != days)
+            {
+                for (int i = 0; i > (days + 1) * -1; i--)
+                {
+                    await this.GetWordOfTheDay(TimeSpan.FromHours(12), start.AddDays(i));
+                }
+                result = await GetWordOfTheDayStats(date, days);
+            }
+            return result;
+        }
     }
 }
