+++ wordle-web/src/views/StatsView.vue
@@ -5,17 +5,23 @@
     </v-card>
     <v-card v-for="play in plays" :key="play.date.toString">
       <v-card color="cyan" style="border: 3px solid red">
-        <v-card-title>Date: {{ play.date }}</v-card-title>
+        <v-card-title>{{ play.date.toDateString() }}</v-card-title>
       </v-card>
       <v-list>
         <v-list-item>
           <v-list-item-title>Games Played: {{ play.numberOfPlays }}</v-list-item-title>
         </v-list-item>
         <v-list-item>
-          <v-list-item-title>Average Score: {{ play.averageAttempts }}</v-list-item-title>
+          <v-list-item-title
+            >Average Attempts:
+            {{ play.avgAttempts == -1 ? '' : Math.round(play.avgAttempts) }}</v-list-item-title
+          >
         </v-list-item>
         <v-list-item>
-          <v-list-item-title>Average Time: {{ play.averageTime }}</v-list-item-title>
+          <v-list-item-title
+            >Average Time:
+            {{ play.avgTime == -1 ? '' : Math.round(play.avgTime) + ' seconds' }}</v-list-item-title
+          >
         </v-list-item>
       </v-list>
     </v-card>
@@ -31,10 +37,11 @@ import { inject, ref } from 'vue'
 const plays = ref<WordOfTheDayStats[]>([])
 const playerService = inject(Services.PlayerService) as PlayerService
 
-Axios.get('/Play/GetLastTenDates?playerId=' + playerService.player.playerId).then((response) => {
+Axios.get('/Word/GetLastTenDates?playerId=' + playerService.player.playerId).then((response) => {
   for (let stat of response.data as WordOfTheDayStats[]) {
     stat.date = new Date(stat.date)
   }
   plays.value = response.data as WordOfTheDayStats[]
+  console.log(plays.value)
 })
 </script>
